#9573. Replication correct handling DeleteRule=Deny.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@13078 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ frameworks/common-ui/src/main/java/ish/oncourse/ui/pages/Courses.java
@@ -73,7 +73,7 @@ public class Courses {
 	@Property
 	private List<Site> mapSites;
 	@Property
-	private Map<Integer, Float> focusesForMapSites;
+	private Map<Long, Float> focusesForMapSites;
 	@InjectComponent
 	private Zone coursesZone;
 	@InjectComponent
@@ -122,7 +122,7 @@ public class Courses {
 	private void setupMapSites() {
 		mapSites = new ArrayList<Site>();
 		if (hasAnyFormValuesForFocus()) {
-			focusesForMapSites = new HashMap<Integer, Float>();
+			focusesForMapSites = new HashMap<Long, Float>();
 		}
 		Double[] locationPoints = { null, null };
 		if (searchParams != null && searchParams.containsKey(SearchParam.near)) {
